feat: Add minimum income to loan products, update loan application validation and implement My Application page

+++ src/main/java/bg/credihub/model/entities/LoanApplication.java
@@ -33,7 +33,7 @@ public class LoanApplication {
     @Column(nullable = false)
     private Integer periodMonths;
     @Column(nullable = false)
-    private Integer monthlyIncome;
+    private BigDecimal monthlyIncome;
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private LoanPurpose loanPurpose;
